Moved backend code to viewmodel and bind approrpiately.
